+++ pyragcore/__init__.py
@@ -1,11 +1,19 @@
-# pyragcore/__init__.py
+#pyragcore/__init__.py
 from pyragcore.pipeline.base_pipeline import BasePipeline
 from pyragcore.embeddings.embedder import Embedder
 from pyragcore.retrieval.vector_store import VectorStore
 from pyragcore.retrieval.retriver import Retriever
 from pyragcore.llm.responder import Responder
-from pyragcore.ingestion.base_loader import BaseLoader
-from pyragcore.ingestion.base_chunker import BaseChunker
+
+# interfaces
+from pyragcore.interfaces.base_loader import BaseLoader
+from pyragcore.interfaces.base_chunker import BaseChunker
+from pyragcore.interfaces.base_embedder import BaseEmbedder
+from pyragcore.interfaces.base_vector_store import BaseVectorStore
+from pyragcore.interfaces.base_llm import BaseLLM
+from pyragcore.interfaces.base_retriever import BaseRetriever
+
+# exceptions
 from pyragcore.exceptions import (
     BotRagException,
     EmbeddingException,
@@ -14,15 +22,21 @@ from pyragcore.exceptions import (
     ModelNotFoundException,
 )
 
-
 __all__ = [
+    # concrete classes
     "BasePipeline",
     "Embedder",
     "VectorStore",
     "Retriever",
     "Responder",
+    # interfaces
     "BaseLoader",
     "BaseChunker",
+    "BaseEmbedder",
+    "BaseVectorStore",
+    "BaseLLM",
+    "BaseRetriever",
+    # exceptions
     "BotRagException",
     "EmbeddingException",
     "RetrievalException",
@@ -30,7 +44,5 @@ __all__ = [
     "ModelNotFoundException",
 ]
 
-
-
-__version__ = "0.1.8"
+__version__ = "0.2.0"
 __author__ = "Vlad Digori"