+++ pyragcore/retrieval/vector_store.py
@@ -1,10 +1,12 @@
+#pyragcore/retrieval/vector_store.py
 import faiss
 import pickle
 import os
 from typing import Optional
 from tqdm import tqdm
 from pyragcore.exceptions import VectorStoreException
-class VectorStore:
+from pyragcore.interfaces.base_vector_store import BaseVectorStore
+class VectorStore(BaseVectorStore):
     def __init__(self,dim:int,persist_path:str,index_type:str="flat",metric:str="l2",normalizr:bool=True,autosave:bool=True,load_if_exist:bool=True):
         self.dim=dim
         self.persist_path=persist_path